fix: resolve NPU dtype mismatches in optimizer
Two changes:
1. Use fp32 compute dtype on Ascend NPU instead of bf16. NPU's
   aclnnInplaceLerp requires all operands to have the same dtype; since
   the 0-D scalar hyperparameter tensors are fp32, keeping everything
   fp32 avoids all mixed-dtype errors.

2. Remove hardcoded .bfloat16() cast in muon_step_fused Polar Express.
   Now uses g.dtype so it stays fp32 on NPU and bf16 on CUDA with bf16
   model weights — no behavior change on existing CUDA bf16 training.

+++ nanochat/optim.py
@@ -114,7 +114,7 @@ def muon_step_fused(
     g = stacked_grads.lerp_(momentum_buffer, momentum)
 
     # Polar express
-    X = g.bfloat16()
+    X = g
     X = X / (X.norm(dim=(-2, -1), keepdim=True) * 1.01 + 1e-6)
     if g.size(-2) > g.size(-1): # Tall matrix
         for a, b, c in polar_express_coeffs[:ns_steps]:
